+++ src/scrapper/rsc/helper.py
@@ -1,0 +1,17 @@
+import os
+
+
+def cls():
+    os.system('cls' if os.name=='nt' else 'clear')
+
+
+
+'''
+    Make Script non executable
+'''
+def main():
+    pass
+
+
+if __name__ == "__main__":
+    main()
